backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
supabase>=2.4.0
google-generativeai>=0.7.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
aiofiles>=23.2.1
pytest>=8.2.0
